# docs/openapi.yml
openapi: 3.1.0
info:
  title: "Maroa.ai API"
  version: 2.3.0
  description: "AI marketing automation SaaS — Inngest + Express backend. See CLAUDE.md for architecture."
servers:
  - url: https://maroa-api-production.up.railway.app
    description: Production
  - url: http://localhost:3000
    description: "Local dev"
...
paths:
...
  /api/lead-magnets/generate:
    post:
      tags:
        - api
      summary: "POST /api/lead-magnets/generate"
      description: "Authenticated dashboard API (POST /api/lead-magnets/generate). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
  /api/lead-magnets/:userId:
    get:
      tags:
        - api
      summary: "GET /api/lead-magnets/:userId"
      description: "Authenticated dashboard API (GET /api/lead-magnets/generate). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/launch/create:
    post:
      tags:
        - api
      summary: "POST /api/launch/create"
      description: "Authenticated dashboard API (POST /api/launch/create). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
  /api/launch/:userId:
    get:
      tags:
        - api
      summary: "GET /api/launch/:userId"
      description: "Authenticated dashboard API (GET /api/launch/create). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/ideas/generate:
    post:
      tags:
        - api
      summary: "POST /api/ideas/generate"
      description: "Authenticated dashboard API (POST /api/ideas/generate). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
  /api/ideas/:userId:
    get:
      tags:
        - api
      summary: "GET /api/ideas/:userId"
      description: "Authenticated dashboard API (GET /api/ideas/generate). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/ideas/:ideaId:
    patch:
      tags:
        - api
      summary: "PATCH /api/ideas/:ideaId"
      description: "Authenticated dashboard API (PATCH /api/ideas/generate). Requires Supabase JWT unless documented as public."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    WebhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
    MetricsToken:
      type: apiKey
      in: header
      name: x-metrics-token
      description: "METRICS_SCRAPE_TOKEN or ORCHESTRATOR_SECRET; Bearer auth also accepted"
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: "Machine-readable error code (e.g. VALIDATION_ERROR, RATE_LIMITED, COST_CAP_REACHED)"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
              nullable: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
...